test: consolidate and refactor typostats test suite
Consolidates the fragmented, tiny test files `test_typostats_coverage_gap.py`, `test_typostats_directory.py`, `test_typostats_interactive.py`, and `test_typostats_main.py` into `tests/test_typostats.py` to reduce domain fragmentation and improve discoverability.

Refactors compound test cases into atomic, descriptively-named unit tests, and completely strips explanatory narrative inline/block comments to align with repository test hygiene rules. All 1,161 tests pass successfully and 100% coverage of typostats.py is preserved.
