--- tests/test_typostats_interactive.py
@@ -1,47 +1,0 @@
-import io
-import sys
-from unittest.mock import patch, MagicMock
-import pytest
-from typostats import main, _STDIN_CACHE
-
-def test_typostats_interactive_terminal_scans_cwd(tmp_path, monkeypatch):
-    # Set the temporary directory as the current working directory
-    monkeypatch.chdir(tmp_path)
-
-    # Create a typo file in this temporary directory
-    file1 = tmp_path / "typos.txt"
-    file1.write_text("teh -> the\n")
-
-    # Clear STDIN cache
-    monkeypatch.setattr("typostats._STDIN_CACHE", None)
-
-    # Mock sys.stdin.isatty to return True, and pass no file arguments
-    with patch("sys.stdin.isatty", return_value=True), \
-         patch("sys.argv", ["typostats.py", "--quiet", "--output", str(tmp_path / "report.json"), "--all"]):
-        main()
-
-    # The report should have been generated by scanning the current directory (which contains typos.txt)
-    report_file = tmp_path / "report.json"
-    assert report_file.exists()
-    content = report_file.read_text()
-    assert '"typo": "eh"' in content or "eh" in content
-
-
-def test_typostats_non_interactive_terminal_reads_stdin(tmp_path, monkeypatch):
-    # Clear STDIN cache
-    monkeypatch.setattr("typostats._STDIN_CACHE", None)
-
-    # Mock stdin content
-    mock_stdin = io.StringIO("recived -> received\n")
-
-    # Mock sys.stdin.isatty to return False, which is the standard non-interactive/piped state
-    with patch("sys.stdin.isatty", return_value=False), \
-         patch("sys.stdin", mock_stdin), \
-         patch("sys.argv", ["typostats.py", "--quiet", "--output", str(tmp_path / "report.json"), "--all"]):
-        main()
-
-    # The report should be generated from stdin
-    report_file = tmp_path / "report.json"
-    assert report_file.exists()
-    content = report_file.read_text()
-    assert '"typo": "i"' in content or "i" in content
